internal/spice: unify branch traversals
The branch graph traversals end up loading the same data from disk
several times to try to look up or down the stack.

This introduces a BranchGraph abstraction that loads that data once
and provides read-only access to the graph as iterables.

For backwards compatiblity, the old methods are still available for now.
We may remove them in the future and pass the graph around.

The graph can also be cached using a hash of the store's Git hash,
so the same process can re-use the graph if the store hasn't changed.
(This is not implemented yet.)