fix(log): Fix regression in deleted branch detection (#772)
Fixes a regression introduced in a prior commit
where 'gs ls' would panic if a branch was deleted out of band.
To fix the issue, switch to using the branch graph
which will get rid of out-of-band deleted branches during load.

No changelog entry needed as the regression wasn't released to users
yet.