Add 'spice.submit.updateOnly' config to set --update-only by default (#776)
The `--update-only` flag is useful and some users would like that to be
the default behavior. We can make this behavior configurable. We follow
the pattern currently used by `Draft` option. I.e. we create a new,
hidden, config flag called `UpdateOnlyDefault` and use this to help set
the now negatable value of the `UpdateOnly` flag.

Resolves #775

---------

Co-authored-by: Abhinav Gupta <mail@abhinavg.net>