doc: add Get Started subsections to sidebar (#794)
These were originally kept out of the sidebar
because I expected them to have verbose titles
and wanted more flexibility with content and subsections.

I think it's fine to keep them in the sidebar now
as the expected UX of the site is now to have subsections
clearly visible in the sidebar.

Resolves #792