fix(submit): capture and surface stdout on push
If a pre-push hook fails and prints messages to stdout and stderr,
we currently only surface stderr in the error message.
This change captures stdout as well and includes it in the error.

Note that the message is surfaced only when the command fails
(except in --verbose mode which logs them both live).

There may be need for an option in the future
where users want to always surface stdout/stderr directly,
rather than logging it/capturing it for errors only.

Resolves: #808