branchPrompter: take more advantage of branch graph
branch graph already has all the information the store.ListBranches
and LookupBranch calls were being used for.
So load the graph once and just roll with it.
(Branch graph did not exist when this code was originally written.)

Taking this approach also means that for repo.LocalBranches,
we can process the output in the iterator directly
rather than taking multiple passes over the entire list.