feat: add 'gs commit fixup' command
Implements a new 'gs commit fixup' command
that acts like --amend for a commit down the stack.

By default, it presents a nice interactive prompt
for selecting a downstack commit to amend.
You can also specify the commit hash directly,
but that leads to an expensive search across all branches.
I don't have an obvious alternative to that yet.

Up until the final rebase and upstack restack,
the command operates without a working tree.
This requires at least Git 2.45 for this to work.

This is a first step to implementing tree-less operations.

In the future, I expect that restack operations
will be based on the same underlying mechanism.

The new commit will re-use metadata from the original commit.
We don't currently offer a means to alter the commit message.
I think doing that with commit-tree will require
re-creating the 'git commit' editing experience,
which is out of scope for this initial implementation.

The staged changes are squashed into the target commit.
This can be made optional if necessary.

---

TODO:

- [ ] Play with this a bit more to see if the UX is good enough
- [ ] Tests
- [ ] Documentation on the website