rebase: use -z flag to handle special characters in file paths
`git ls-files` will escape file paths with special characters in them.
We were only splitting on newline, which isn't good enough.

Use the -z flag, and split on null bytes instead
to obviate any escaping concerns.

Also, when logging the list of files, escape paths for readability
so if there are files literally " ", they are quoted in the log output.

Refs #849