docs: Provide llms.txt and llms-full.txt (#858)
Uses the `mkdocs-llmstxt` plugin to generate structured documentation
files (`llms.txt` and `llms-full.txt`) that LLMs can easily parse and
reference.

Currently, these files expose most of the same content as the main
documentation site, except for the changelog. 

To ensure code snippets wrapped in SVGs (created by freeze) are included
in the output, we also wrap them in a hidden `<code>` tag. This improves
accessibility for both screen readers and allows the plugin generating
the `llms.txt` file to include it in its output.

This only a first minimal implementation. Further iterations may be
required to ensure cleaner markdown of the documentation pages.

Closes #827

---------

Co-authored-by: Abhinav Gupta <mail@abhinavg.net>