branch split: Allow using original name for mid commits
Today, 'gs branch split' requires branch names for intermediate commits
to be unique and different from the original branch name.
This is good for when the split commit is intended to precede
the original branch:

    o (main)                       o (main)
     \                              \
      A --> B (feature)    =>        A (feature-pre)
                                      \
                                       B (feature)

However, this does not work for cases where the split
should take over the original branch, such as:

    o (main)                       o (main)
     \                              \
      A --> B (feature)    =>        A (feature)
                                      \
                                       B (feature-post)

This commit adds support for reusing the original branch name
for an intermediate commit during a branch split.

This works for interactive and non-interactive modes.

- For interactive mode, if any of the intermediate commits
  request the original branch name,
  a new field is presented to name the original HEAD commit.
- For non-interactive mode, if any of the split points
  reuse the original branch name,
  it is required that a split point for the HEAD commit
  is also provided with a different name.

Resolves #411