feat(log): opt-in to show CR status (#866)
Adds a --cr-status/-S flag to `gs ls` and `gs ll`
that will query the underlying Forge for the statuses of all open CRs
and display them in the resulting log.

The option can be made default with the spice.log.crStatus option
but it's not recommended as each call is a network request.
Also works in JSON.

In the future, if we teach the forges about approval status,
mergability, etc., we can show that information there as well.

---------

Co-authored-by: Abhinav Gupta <mail@abhinavg.net>