submit: Allow cycling through commit titles with up/down (#906)
When submitting a pull request with `gs * submit`,
the default PR title is based on the first commit in a branch.
This is usually fine but it's not infrequent for someone to have the
correct title at the top instead of the bottom.

Add the ability for users to select titles from other commits
in an unobtrusive way by simply pressing up/down in the title prompt.
Commit titles will be made available in chronological order.

When this option is present, a simple indicator suggesting up/down
arrows
is also rendered as part of the description.

Resolves #896 

---

Demo from original PR:


https://github.com/user-attachments/assets/72949a1d-2496-484b-b5db-7c521e7b208b

Demo following new changes:


![submit-title-select-demo](https://github.com/user-attachments/assets/ba1810d3-a827-4ed9-94da-5d28c9109f34)

---------

Co-authored-by: Abhinav Gupta <mail@abhinavg.net>