submit: Add --assign flag and assignees configuration (#951)
Add support for assigning change requests to users during submission
with an `-a`/`--asign` flag for all submit commands.

The accompanying `spice.submit.assignees` configuration option
allows setting default assignees.

LLM assistance:
The first version of this change was made with ChatGPT Codex.
The result was iterated upon significantly locally,
both manually and with the help of Claude.