test(forge): DRY up github and gitlab integration tests (#973)
The integration tests for github and gitlab are a slog to write and
maintain
because it's just so much copy-pasta.
Both sets of tests are almost exactly the same,
and even within each forge, the individual tests themselves
have a lot of copy paste for the update vs non-update scenarios.

DRY it up into forgetest.RunIntegration that will run acceptance
scenarios on a given forge, providing VCR functionality, etc.