test: verify shamhub behavior compliance (#974)
Take advantage of the generalized integration tests
which verify behavior of github and gitlab
to also verify that our test shamhub implementation
complies with the same behavior.

This is necessary since we use shamhub
for all of git-spice's own integration tests
so any divergence in behavior would be problematic.

This found minor discrepancies in ShamHub behavior that have been fixed.