forge: Assignee and Reviewers are append-only operations (#976)
Reviewers and Assignees should be add-only operations.
This verifies that with a test.

Unfortunately, I was unable to test this for GitLab
because as far as I can tell,
free plan does not support multiple reviewers.
If anyone with a paid plan wants to help test this,
please reach out and I can add you to the test repo
to do this.

Resolves #967